requirements.txt
fastapi[all]
uvicorn[standard]
sqlmodel>=0.0.8
pandas
scikit-learn
mlflow==2.6.0
boto3
jupyterlab
pymysql
python-dotenv
psycopg2-binary